requirements.txt
kafka-python
pyspark
 
